+++ main_test.go
@@ -3,11 +3,9 @@ package main
 import (
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
-	"github.com/googlecloudplatform/gcsfuse/internal/config"
 	mountpkg "github.com/googlecloudplatform/gcsfuse/internal/mount"
 	. "github.com/jacobsa/ogletest"
 )
@@ -73,64 +71,4 @@ func (t *MainTest) TestGetUserAgentWhenMetadataImageTypeEnvVarSetAndAppNameNotSe
 	expectedUserAgent := strings.TrimSpace(fmt.Sprintf("gcsfuse/%s (GPN:gcsfuse-DLVM)", getVersion()))
 
 	ExpectEq(expectedUserAgent, userAgent)
-}
-
-func (t *MainTest) TestOverrideLoggingFlags_WithNonEmptyLogConfigs() {
-	flags := &flagStorage{
-		LogFile:   "a.txt",
-		LogFormat: "json",
-		DebugFuse: true,
-		DebugGCS:  false,
-	}
-	mountConfig := &config.MountConfig{}
-	mountConfig.LogConfig = config.LogConfig{
-		Severity:  config.ERROR,
-		LogFile:   "/tmp/hello.txt",
-		LogFormat: "text",
-	}
-	mountConfig.WriteConfig = config.WriteConfig{
-		CreateEmptyFile: true,
-	}
-
-	overrideWithLoggingFlags(mountConfig, flags)
-
-	AssertEq(mountConfig.LogFormat, "text")
-	AssertEq(mountConfig.LogFile, "/tmp/hello.txt")
-	AssertEq(mountConfig.LogConfig.Severity, config.TRACE)
-}
-
-func (t *MainTest) TestOverrideLoggingFlags_WithEmptyLogConfigs() {
-	flags := &flagStorage{
-		LogFile:   "a.txt",
-		LogFormat: "json",
-	}
-	mountConfig := &config.MountConfig{}
-	mountConfig.LogConfig = config.LogConfig{
-		Severity:  config.INFO,
-		LogFile:   "",
-		LogFormat: "",
-	}
-	mountConfig.WriteConfig = config.WriteConfig{
-		CreateEmptyFile: true,
-	}
-
-	overrideWithLoggingFlags(mountConfig, flags)
-
-	AssertEq(mountConfig.LogFormat, "json")
-	AssertEq(mountConfig.LogFile, "a.txt")
-	AssertEq(mountConfig.Severity, config.INFO)
-}
-
-func (t *MainTest) TestResolveConfigFilePaths() {
-	mountConfig := &config.MountConfig{}
-	mountConfig.LogConfig = config.LogConfig{
-		LogFile: "~/test.txt",
-	}
-
-	err := resolveConfigFilePaths(mountConfig)
-
-	AssertEq(nil, err)
-	homeDir, err := os.UserHomeDir()
-	AssertEq(nil, err)
-	ExpectEq(filepath.Join(homeDir, "test.txt"), mountConfig.LogFile)
 }
